fix(trusty-mpm): statusline shows tmux session name instead of worktree UUID branch
Closes #2031
